Added token to creating a new user
