+++ server.js
@@ -2,7 +2,6 @@ const express = require('express');
 const mongoose = require('mongoose');
 const helmet = require('helmet');
 const cors = require('cors');
-const jwt = require('jsonwebtoken');
 require('dotenv').config();
 
 // Routes
@@ -28,26 +27,6 @@ server.use(express.json());
 server.use(cors());
 server.use('/api', noteRouter);
 server.use('/api', userRouter);
-
-// Helper Functions
-const getTokenForUser = userObject => {
-    return jwt.sign(userObject, process.env.TOKEN_SECRET, { expiresIn: '1h' })
-}
-
-const validateToken = (req, res, next) => {
-    const token = req.headers.authorization;
-    if (!token) {
-        res.status(422).json({ Error: 'No token found' })
-    } else {
-        jwt.verify(token, process.env.TOKEN_SECRET, (err, decoded) => {
-            if (err) {
-                res.status(401).json({ Error: "Token invalid", message: err });
-            } else {
-                next();
-            }
-        })
-    }
-}
 
 // Initial GET
 server.get('/', (req, res) => {
